backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
motor==3.7.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
python-dotenv==1.1.1
